Fixed integer parsing error. Code compiles now
Changed Integers to Doubles and now everything works.
Also made some optimization edits.

+++ Question4/src/main/java/com/revature/question4/FemaleEmployment.java
@@ -3,6 +3,7 @@ package com.revature.question4;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
@@ -30,7 +31,7 @@ public class FemaleEmployment extends Configured implements Tool{
 		job.setReducerClass(Reduce1.class);
 
 		job.setOutputKeyClass(Text.class);
-		job.setOutputValueClass(IntWritable.class);
+		job.setOutputValueClass(DoubleWritable.class);
 
 		boolean success = job.waitForCompletion(true);
 		return success ? 0 : 1;
