+++ Question4/src/main/java/com/revature/question4/Map1.java
@@ -2,13 +2,14 @@ package com.revature.question4;
 
 import java.io.IOException;
 
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 import org.apache.hadoop.mapreduce.Mapper.Context;
 
-public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
+public class Map1 extends Mapper<LongWritable, Text, Text, DoubleWritable>{
 	@Override
 	public void map(LongWritable key, Text text, Context context) throws IOException, InterruptedException{
 		String line = text.toString();
@@ -22,10 +23,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("Female Employment % in Agriculture: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("Female Employment % in Agriculture: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("Female Employment % in Agriculture: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("Female Employment % in Agriculture: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
@@ -39,10 +40,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("% of Female Employers: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("% of Female Employers: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("% of Female Employers: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("% of Female Employers: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
@@ -56,10 +57,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("% of Own-Account Female Workers: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("% of Own-Account Female Workers: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("% of Own-Account Female Workers: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("% of Own-Account Female Workers: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
@@ -73,10 +74,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("% of Self-Employed Female Workers: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("% of Self-Employed Female Workers: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("% of Self-Employed Female Workers: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("% of Self-Employed Female Workers: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
@@ -90,10 +91,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("% of Wage/Salaried Female Workers: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("% of Wage/Salaried Female Workers: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("% of Wage/Salaried Female Workers: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("% of Wage/Salaried Female Workers: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
@@ -107,10 +108,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("Female Employment % in Industry: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("Female Employment % in Industry: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("Female Employment % in Industry: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("Female Employment % in Industry: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
@@ -124,10 +125,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("Female Employment % in Services: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("Female Employment % in Services: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("Female Employment % in Services: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("Female Employment % in Services: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
@@ -141,10 +142,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("Employment to Population Ratio, Female: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("Employment to Population Ratio, Female: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("Employment to Population Ratio, Female: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("Employment to Population Ratio, Female: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
