+++ Question5/src/main/java/com/revature/question5/CustomQuestion.java
@@ -3,6 +3,7 @@ package com.revature.question5;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
@@ -33,7 +34,7 @@ public class CustomQuestion extends Configured implements Tool{
 		job.setReducerClass(Reduce1.class);
 
 		job.setOutputKeyClass(Text.class);
-		job.setOutputValueClass(IntWritable.class);
+		job.setOutputValueClass(DoubleWritable.class);
 
 		boolean success = job.waitForCompletion(true);
 		return success ? 0 : 1;
